+++ weblate/addons/models.py
@@ -5,7 +5,7 @@
 from __future__ import annotations
 
 import logging
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, overload
 
 import sentry_sdk
 from appconf import AppConf
@@ -328,15 +328,54 @@ def execute_addon_event(
                 )
 
 
+@overload
 def handle_addon_event(
     event: AddonEvent,
-    method: str | Callable,
-    args: tuple | None = None,
+    method: str,
+    args: tuple,
     *,
-    component: Component | None = None,
-    translation: Translation | None = None,
+    component: Component,
+    translation: None = None,
     addon_queryset: AddonQuerySet | None = None,
     auto_scope: bool = False,
+) -> None: ...
+
+
+@overload
+def handle_addon_event(
+    event: AddonEvent,
+    method: str,
+    args: tuple,
+    *,
+    component: None = None,
+    translation: Translation,
+    addon_queryset: AddonQuerySet | None = None,
+    auto_scope: bool = False,
+) -> None: ...
+
+
+@overload
+def handle_addon_event(
+    event: AddonEvent,
+    method: Callable,
+    args: None = None,
+    *,
+    component: None = None,
+    translation: None = None,
+    addon_queryset: AddonQuerySet | None,
+    auto_scope: bool,
+) -> None: ...
+
+
+def handle_addon_event(
+    event,
+    method,
+    args=None,
+    *,
+    component=None,
+    translation=None,
+    addon_queryset=None,
+    auto_scope=False,
 ) -> None:
     # Scope is used for logging
     scope = translation or component
@@ -345,13 +384,16 @@ def handle_addon_event(
     if component is None and translation is not None:
         component = translation.component
 
+    if component is None and not auto_scope:
+        raise ValueError("Missing event scope!")
+
     # EVENT_DAILY uses custom queryset because it is not triggered from the
     # object scope
     if addon_queryset is None:
         addon_queryset = Addon.objects.filter_event(component, event)
 
     for addon in addon_queryset:
-        if not auto_scope:
+        if scope is not None:
             execute_addon_event(addon, component, scope, event, method, args)
         else:
             components: Iterable[Component]
@@ -369,7 +411,7 @@ def handle_addon_event(
 
 
 @receiver(vcs_pre_push)
-def pre_push(sender, component, **kwargs) -> None:
+def pre_push(sender, component: Component, **kwargs) -> None:
     handle_addon_event(
         AddonEvent.EVENT_PRE_PUSH,
         "pre_push",
@@ -379,7 +421,7 @@ def pre_push(sender, component, **kwargs) -> None:
 
 
 @receiver(vcs_post_push)
-def post_push(sender, component, **kwargs) -> None:
+def post_push(sender, component: Component, **kwargs) -> None:
     handle_addon_event(
         AddonEvent.EVENT_POST_PUSH,
         "post_push",
@@ -391,7 +433,7 @@ def post_push(sender, component, **kwargs) -> None:
 @receiver(vcs_post_update)
 def post_update(
     sender,
-    component,
+    component: Component,
     previous_head: str,
     child: bool = False,
     skip_push: bool = False,
@@ -406,7 +448,7 @@ def post_update(
 
 
 @receiver(component_post_update)
-def component_update(sender, component, **kwargs) -> None:
+def component_update(sender, component: Component, **kwargs) -> None:
     handle_addon_event(
         AddonEvent.EVENT_COMPONENT_UPDATE,
         "component_update",
@@ -416,7 +458,7 @@ def component_update(sender, component, **kwargs) -> None:
 
 
 @receiver(vcs_pre_update)
-def pre_update(sender, component, **kwargs) -> None:
+def pre_update(sender, component: Component, **kwargs) -> None:
     handle_addon_event(
         AddonEvent.EVENT_PRE_UPDATE,
         "pre_update",
@@ -426,7 +468,7 @@ def pre_update(sender, component, **kwargs) -> None:
 
 
 @receiver(vcs_pre_commit)
-def pre_commit(sender, translation, author, **kwargs) -> None:
+def pre_commit(sender, translation: Translation, author, **kwargs) -> None:
     handle_addon_event(
         AddonEvent.EVENT_PRE_COMMIT,
         "pre_commit",
@@ -436,7 +478,7 @@ def pre_commit(sender, translation, author, **kwargs) -> None:
 
 
 @receiver(vcs_post_commit)
-def post_commit(sender, component, **kwargs) -> None:
+def post_commit(sender, component: Component, **kwargs) -> None:
     handle_addon_event(
         AddonEvent.EVENT_POST_COMMIT,
         "post_commit",
@@ -446,7 +488,7 @@ def post_commit(sender, component, **kwargs) -> None:
 
 
 @receiver(translation_post_add)
-def post_add(sender, translation, **kwargs) -> None:
+def post_add(sender, translation: Translation, **kwargs) -> None:
     handle_addon_event(
         AddonEvent.EVENT_POST_ADD,
         "post_add",
@@ -456,7 +498,7 @@ def post_add(sender, translation, **kwargs) -> None:
 
 
 @receiver(unit_pre_create)
-def unit_pre_create_handler(sender, unit, **kwargs) -> None:
+def unit_pre_create_handler(sender, unit: Unit, **kwargs) -> None:
     handle_addon_event(
         AddonEvent.EVENT_UNIT_PRE_CREATE,
         "unit_pre_create",
@@ -467,7 +509,7 @@ def unit_pre_create_handler(sender, unit, **kwargs) -> None:
 
 @receiver(post_save, sender=Unit)
 @disable_for_loaddata
-def unit_post_save_handler(sender, instance, created, **kwargs) -> None:
+def unit_post_save_handler(sender, instance: Unit, created, **kwargs) -> None:
     handle_addon_event(
         AddonEvent.EVENT_UNIT_POST_SAVE,
         "unit_post_save",
@@ -477,7 +519,7 @@ def unit_post_save_handler(sender, instance, created, **kwargs) -> None:
 
 
 @receiver(store_post_load)
-def store_post_load_handler(sender, translation, store, **kwargs) -> None:
+def store_post_load_handler(sender, translation: Translation, store, **kwargs) -> None:
     handle_addon_event(
         AddonEvent.EVENT_STORE_POST_LOAD,
         "store_post_load",